cleaned up gql type defs
