+++ src/apollo/server/schema.js
@@ -1,8 +1,17 @@
 import { makeExecutableSchema } from 'graphql-tools';
-import { typeDefs } from './type-defs';
 import { resolvers } from './resolvers';
+// Types
+import { typeDefs as LocationTypes } from './type-defs/location.js';
+import { typeDefs as OnlineResourceTypes } from './type-defs/onlineResource.js';
+import { typeDefs as TaxonomyTypes } from './type-defs/taxonomy.js';
+import { typeDefs as SharedTypes } from './type-defs/shared.js';
 
 export const schema = makeExecutableSchema({
-  typeDefs,
+  typeDefs: [ 
+    SharedTypes,
+    LocationTypes, 
+    OnlineResourceTypes, 
+    TaxonomyTypes,
+  ],
   resolvers,
 })
