+++ src/apollo/server/type-defs/taxonomy.js
@@ -1,0 +1,21 @@
+import gql from 'graphql-tag';
+
+export const typeDefs = gql`
+  type Vocab {
+    id: ID!
+    name: String!
+    terms: [Term]
+  }
+
+  type Term {
+    id: ID!
+    name: String!
+    children: [Term]
+  }
+
+  input TermsFilter {
+    id: String!
+    terms: [String]!
+    operator: String!
+  }
+`;
